add EAN params customerSessionId, customerIpAddress, and customerUserAgent. improve date selection and checking. general cleanup

+++ ota-ios/ota-ios/Models/Internal/ChildTraveler.h
@@ -24,7 +24,7 @@ typedef NS_ENUM(NSUInteger, CHILD_TRAVELER_ID) {
 
 + (ChildTraveler *)childTravelerForId:(CHILD_TRAVELER_ID)childTravelerId;
 + (NSArray *)childTravelers;
-+ (NSUInteger)numberOfKids;
++ (int)numberOfKids;
 + (NSInteger)addChildTraveler;
 + (NSInteger)removeLastChildTraveler;
 + (BOOL)moreKidsOk;
